It's nearing completion. Here's a test version, and it was a success, it's much much faster: http://pastebin.com/m1KCzjdk
It just needs to be sourced and it will bind to <TAB>.
I'm now going to use it for my day-to-day needs. Looking for bugs and
things I forgot. If you find something missing let me know!
Priority of media files per command is already in.
Sorting by time is out. It's just not feasible with the way the file
list is acquired (just a glob).
CTRL+C doesn't work on the menu for some reason. Only Escape. This
should be solved by making a separate script file instead of a function.
I'll try that next time.
Other Todo's are bringing directories to the top and completing commands
after sudo and setsid etc. and completing files after =. Maybe add keys
for left and right to go into other directories.