https://bbs.archlinux.org/viewtopic.php?pid=1222146#p1222146 => http://pastebin.com/aNWpV4mV
This time I put it in a script. That way CTRL+C works, but the bind -x
line becomes a bit more complex and all output has to go to STDERR.
There's a huge speedup this time. I took out the code that stat's every
file. Only when a file is displayed in the menu is the size calculated
and checked whether it's a directory.
For cd it will still stat every file though, to put all the directories
first.
That means in an uncached directory of 2000 files, it will start
displaying in 200 milliseconds instead of 7 seconds.
I added F2: change the command without exiting (selected file is
preserved).
and F3: sort by time (it's a bit limited, but enough in my opnion).
Left and Right work, but when there's only one item, it's auto selected.
Also on Left the directory you came from isn't selected. I think I may
be able to fix both, but I haven't thought about how to yet.
I made it so you have to press Right twice. The first time it will show
some information about the file or directory. This can easily be removed
if you don't like it.
I played around a bit with the menu colors. Directories will now be
underlined and matching media files will be bright blue.
The globbing got stuck on [abcd] patterns but that should be fixed now.