https://bbs.archlinux.org/viewtopic.php?pid=1224295#p1224295
I put the info text for the selected file on F1. Right goes on at once.
I replaced directory underline with making the "size" field stand out
more using unicode blocks: ▓▓▓▓
I got CTRL+C working by using a subshell for just `read`.
There was a bug with LC_ALL that messed with the prompt on really long
lines. Now it uses LC_CTYPE. You should use a fresh shell to see the
change.
I also forgot that the python script is a console dmenu. A function
can't be called in an external script like that, so it has to be a
separate file. I made a new small console dmenu in bash. You can move
left and right in the prompt. F2 accepts the line as-is. Tab will fill
in the current item. Spaces make an "AND"-like grep for each word.