https://bbs.archlinux.org/viewtopic.php?pid=1228152#p1228152
Yeah, I put it on a timer now. 0.1ms seemed to be enough, but I put it
on 2ms just to be sure.
So if you press <tab> on an empty line now, you can complete a file,
afterwards it will put the cursor at the start of the line.
I also made it so that command completions from scratch like with F2 and
deleting all characters makes it so that it will prompt for the first
key.
Because a list of all commands sometimes gives kind of an annoying
delay.
I also added a help bar at the bottom. You can remove it by running the
bind -x command with "nohelp" as an argument:
bind -x '"\t":kb_main nohelp'
It's in the examples at the bottom of the file too.